requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.14
pymongo==4.6.1
python-dotenv==1.0.1
itsdangerous==2.2.0
